[HOOZ-16] Configure application SMTP settings
This commit exposes the SMTP settings for the production environment to
environmentn variables. These can be set to point to sendgrid for
production and mailtrap.io for staging and qa.

In development mode the letter opener gem had been installed, to catch
and emails being generated and open them up in a new browser window
for inspection.

The default from address in the `ApplicationMailer` has been updated
to 'info@hoozi.com` which generated mailers will inherit from.