Add Homeowner login screen content into the db
This commit sets up the homeowners login screen to use different
marketing content as the mockups require. The content now lives in the
database and is pulled out randomly to be shown on the login screen,
but if no records exist then a set of configurable defaults will be
used. Eventually there will need to be a CRUD section added to the admin
to manage the different content that can be shown here.

- Add a service to encapsulate the change
- Randomly choose one of the login content records to show
- Add default login content into the seeds
- Use default content for the login screen if nothing has been specified